add unit and FillDestroyFile support

+++ pkg/server/chaosd/disk.go
@@ -34,11 +34,11 @@ type diskAttack struct{}
 
 var DiskAttack AttackType = diskAttack{}
 
-const DDWritePayloadCommand = "dd if=/dev/zero of=%s bs=%s count=%s oflag=dsync"
-const DDReadPayloadCommand = "dd if=%s of=/dev/null bs=%s count=%s iflag=dsync,direct,fullblock"
+const DDWritePayloadCommand = "dd if=/dev/zero of=%s bs=1%s count=%s oflag=dsync"
+const DDReadPayloadCommand = "dd if=%s of=/dev/null bs=1%s count=%s iflag=dsync,direct,fullblock"
 
 func (disk diskAttack) Attack(options core.AttackConfig, env Environment) (err error) {
-	attack := options.(*core.DiskCommand)
+	attack := options.(*core.DiskOption)
 
 	if options.String() == core.DiskFillAction {
 		return disk.attackDiskFill(attack)
@@ -46,13 +46,13 @@ func (disk diskAttack) Attack(options core.AttackConfig, env Environment) (err e
 	return disk.attackDiskPayload(attack)
 }
 
-func (diskAttack) attackDiskPayload(payload *core.DiskCommand) error {
+func (diskAttack) attackDiskPayload(payload *core.DiskOption) error {
 	switch payload.Action {
 	case core.DiskWritePayloadAction:
 		if payload.Path == "" {
 			payload.Path = "/dev/null"
 		}
-		cmd := exec.Command("bash", "-c", fmt.Sprintf(DDWritePayloadCommand, payload.Path, "1M", payload.Size))
+		cmd := exec.Command("bash", "-c", fmt.Sprintf(DDWritePayloadCommand, payload.Path, payload.Unit, payload.Size))
 		output, err := cmd.CombinedOutput()
 
 		if err != nil {
@@ -67,7 +67,7 @@ func (diskAttack) attackDiskPayload(payload *core.DiskCommand) error {
 			log.Error(fmt.Sprintf("payload action: %s", payload.Action), zap.Error(err))
 			return err
 		}
-		cmd := exec.Command("bash", "-c", fmt.Sprintf(DDReadPayloadCommand, payload.Path, "1M", payload.Size))
+		cmd := exec.Command("bash", "-c", fmt.Sprintf(DDReadPayloadCommand, payload.Path, payload.Unit, payload.Size))
 		output, err := cmd.CombinedOutput()
 
 		if err != nil {
@@ -83,10 +83,10 @@ func (diskAttack) attackDiskPayload(payload *core.DiskCommand) error {
 	}
 }
 
-const DDFillCommand = "dd if=/dev/zero of=%s bs=%s count=%s iflag=fullblock"
-const DDFallocateCommand = "fallocate -l %sM %s"
+const DDFillCommand = "dd if=/dev/zero of=%s bs=1%s count=%s iflag=fullblock"
+const DDFallocateCommand = "fallocate -l %s%s %s"
 
-func (diskAttack) attackDiskFill(fill *core.DiskCommand) error {
+func (diskAttack) attackDiskFill(fill *core.DiskOption) error {
 	if fill.Path == "" {
 		tempFile, err := ioutil.TempFile("", "example")
 		if err != nil {
@@ -107,13 +107,16 @@ func (diskAttack) attackDiskFill(fill *core.DiskCommand) error {
 		}
 
 		fill.Path = tempFile.Name()
-		defer func() {
+	}
+
+	defer func() {
+		if fill.FillDestroyFile {
 			err := os.Remove(fill.Path)
 			if err != nil {
-				log.Error(fmt.Sprintf("unexpected err when removing temp file %s", fill.Path), zap.Error(err))
+				log.Error(fmt.Sprintf("unexpected err when removing file %s", fill.Path), zap.Error(err))
 			}
-		}()
-	}
+		}
+	}()
 
 	var cmd *exec.Cmd
 	if fill.Size != "" {
@@ -135,13 +138,14 @@ func (diskAttack) attackDiskFill(fill *core.DiskCommand) error {
 		reservedBlocks := s.Bfree - s.Bavail
 		totalM := uint64(s.Frsize) * (s.Blocks - reservedBlocks) / 1024 / 1024
 		fill.Size = strconv.FormatUint(totalM*percent/100, 10)
+		fill.Unit = "M"
 	}
 
 	if fill.FillByFallocate {
-		cmd = exec.Command("bash", "-c", fmt.Sprintf(DDFallocateCommand, fill.Size, fill.Path))
+		cmd = exec.Command("bash", "-c", fmt.Sprintf(DDFallocateCommand, fill.Size, fill.Unit, fill.Path))
 	} else {
 		//1M means the block size. The bytes size dd read | write is (block size) * (size).
-		cmd = exec.Command("bash", "-c", fmt.Sprintf(DDFillCommand, fill.Path, "1M", fill.Size))
+		cmd = exec.Command("bash", "-c", fmt.Sprintf(DDFillCommand, fill.Path, fill.Unit, fill.Size))
 	}
 
 	output, err := cmd.CombinedOutput()
